Point of parity problem
Mr. Pepper is facing a problem in understanding some information.
He will be able to able to understand the information, if he is able
to find "point of parity".
Your task is to help him find such point. 
(if "point of parity" exists print such point, otherwise print "-1")

"point of parity" is defined as a place (index),which acts as a neutral-point.